Complete error validation tests
